refactor(voice-call): split webhook server and tailscale helpers
(cherry picked from commit 439a7732f4787bf46252986719c818fd8d94bbcf)
