+++ extensions/voice-call/src/tunnel.ts
@@ -1,5 +1,5 @@
 import { spawn } from "node:child_process";
-import { getTailscaleDnsName } from "./webhook.js";
+import { getTailscaleDnsName } from "./webhook/tailscale.js";
 
 /**
  * Tunnel configuration for exposing the webhook server.
